feat(FE): changed doctor chat to swipable version
